Fashion Concierge

+++ app/src/main/java/com/ndg/intel/fashionconcierge/UserAuthActivity.java
@@ -24,12 +24,16 @@ import com.jjoe64.graphview.helper.StaticLabelsFormatter;
 import com.jjoe64.graphview.series.DataPoint;
 import com.jjoe64.graphview.series.LineGraphSeries;
 
+import org.apache.commons.math3.analysis.function.Min;
+import org.apache.commons.math3.stat.correlation.PearsonsCorrelation;
+import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Iterator;
+import java.util.List;
 import java.util.UUID;
 
 
@@ -81,7 +85,7 @@ public class UserAuthActivity extends ActionBarActivity implements SensorEventLi
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
         getMenuInflater().inflate(R.menu.menu_user_auth, menu);
-        return true;
+        return super.onCreateOptionsMenu(menu);
     }
 
     @Override
@@ -91,9 +95,15 @@ public class UserAuthActivity extends ActionBarActivity implements SensorEventLi
         // as you specify a parent activity in AndroidManifest.xml.
         int id = item.getItemId();
 
-        //noinspection SimplifiableIfStatement
-        if (id == R.id.action_settings) {
-            return true;
+        // Handle presses on the action bar items
+        switch (id) {
+            case R.id.action_save_hrm_profile:
+                saveHRMProfile(true);
+                return true;
+            case R.id.action_settings:
+                break;
+            default:
+                break;
         }
 
         return super.onOptionsItemSelected(item);
@@ -103,18 +113,15 @@ public class UserAuthActivity extends ActionBarActivity implements SensorEventLi
     protected void onStart() {
         super.onStart();
         // Register HRM sensor listener
-        mSensorManager.registerListener(this, this.mHeartRateSensor, 3);
-        clearHRMProfile();
+        mSensorManager.registerListener(this, this.mHeartRateSensor, SensorManager.SENSOR_DELAY_FASTEST);
     }
 
     @Override
     protected void onStop() {
         super.onStop();
         // Unregister HRM sensor listener
-        clearHRMProfile();
+        clearAll();
         mSensorManager.unregisterListener(this);
-        mHrmState = HRMSensorState.NOT_READY;
-        timestamp = 0;
     }
 
     private StoreAssociate getStoreAssociate() {
@@ -137,21 +144,21 @@ public class UserAuthActivity extends ActionBarActivity implements SensorEventLi
         userAuthImage.setImageResource(R.drawable.heart);
 
         mGraphView = (GraphView) findViewById(R.id.graph);
-//        mGraphView.getGridLabelRenderer().setHorizontalAxisTitle("time (sec)");
-//        mGraphView.getGridLabelRenderer().setVerticalAxisTitle("bpm");
+ //       mGraphView.getGridLabelRenderer().setHorizontalAxisTitle("time (sec)");
+ //       mGraphView.getGridLabelRenderer().setVerticalAxisTitle("bpm");
 
         mGraphView.getViewport().setXAxisBoundsManual(true);
         mGraphView.getViewport().setMinX(0);
         mGraphView.getViewport().setMaxX(10);
+/*
         mGraphView.getViewport().setYAxisBoundsManual(true);
-        mGraphView.getViewport().setMinY(70);
-        mGraphView.getViewport().setMaxY(100);
-
+        mGraphView.getViewport().setMinY(60);
+        mGraphView.getViewport().setMaxY(120);
+*/
         StaticLabelsFormatter staticLabelsFormatter = new StaticLabelsFormatter(mGraphView);
         staticLabelsFormatter.setHorizontalLabels(new String[] {"0", "1", "2", "3", "4", "5", "6", "7", "8", "9", "10"});
-        staticLabelsFormatter.setVerticalLabels(new String[]{"70", "80", "90", "100"});
+        //staticLabelsFormatter.setVerticalLabels(new String[]{"70", "80", "90", "100"});
         mGraphView.getGridLabelRenderer().setLabelFormatter(staticLabelsFormatter);
-
         mSeries = new LineGraphSeries<DataPoint>();
         mGraphView.addSeries(mSeries);
     }
@@ -160,15 +167,31 @@ public class UserAuthActivity extends ActionBarActivity implements SensorEventLi
         Handler handler = new Handler();
         final Runnable r = new Runnable() {
             public void run() {
+                saveHRMProfile(false);
+                //logHRMProfile();
                 matchUserProfile();
             }
         };
 
-        handler.postDelayed(r, 10000);
+        handler.postDelayed(r, 11000);
+    }
+
+    private void delayedRestartHRM() {
+        Handler handler = new Handler();
+        final SensorEventListener listener = this;
+        final Runnable r = new Runnable() {
+            public void run() {
+                clearAll();
+                mSensorManager.registerListener(listener, mHeartRateSensor,
+                                                SensorManager.SENSOR_DELAY_FASTEST);
+            }
+        };
+
+        handler.postDelayed(r, 5000);
     }
 
     private void matchUserProfile() {
-
+/*
         double savedHRMProfile[] = new double[] {74.28806, 74.80511, 76.07594, 77.69099,
                                                  79.06474, 80.16972, 80.84673, 81.23549,
                                                  81.20487, 80.73627, 80.39274, 80.1198,
@@ -177,13 +200,35 @@ public class UserAuthActivity extends ActionBarActivity implements SensorEventLi
                                                    80.743576, 80.90538, 80.761986, 80.37992,
                                                    79.92841, 79.84282, 79.797615, 79.88893,
                                                    80.02692};
+*/
+        // Obtain stored and current HRM profile
+        List<String> savedHRMProfile = getHRMProfile();
+        List<String> currentHRMProfile = getHRMSeries();
 
-        if (corr(savedHRMProfile, currentHRMProfile) > 0.5) {
-            mHrmState = HRMSensorState.SUCCESS;
-        }
-        else {
+        // Convert to double arrays
+        int szArray = Math.min(savedHRMProfile.size(), currentHRMProfile.size());
+        // If less than 5 probe points the correlation cannot be established
+        if (szArray < 5) {
             mHrmState = HRMSensorState.FAIL;
+            return;
         }
+
+        double dSavedProfile[] = new double[szArray];
+        double dCurrentProfile[] = new double[szArray];
+
+        for (int i=0; i < szArray; i++) {
+            dSavedProfile[i] = Double.parseDouble(savedHRMProfile.get(i));
+            dCurrentProfile[i] = Double.parseDouble(currentHRMProfile.get(i));
+        }
+
+        double confidence = corr(dSavedProfile, dCurrentProfile);
+        displayConfidence(confidence);
+        if (confidence < 0.5) {
+            mHrmState = HRMSensorState.FAIL;
+            return;
+        }
+
+        mHrmState = HRMSensorState.SUCCESS;
     }
 
     /**
@@ -192,14 +237,19 @@ public class UserAuthActivity extends ActionBarActivity implements SensorEventLi
     public double corr(double[] x1, double[] x2)
     {
         //use http://commons.apache.org/proper/commons-math/userguide/stat.html#a1.7_Covariance_and_correlation
+        PearsonsCorrelation pearsonsCorrelation = new PearsonsCorrelation();
+        double corr = pearsonsCorrelation.correlation(x1, x2);
 
-        return 0;
+        return corr;
     }
 
     @Override
     public void onSensorChanged(SensorEvent event) {
-        //Log.i("UserAuthActivity", "sensor event accuracy: " + event.accuracy + " = " + event.values[0]);
+        Log.i("UserAuthActivity", "sensor event accuracy: " + event.accuracy + " = " + event.values[0]);
         switch (mHrmState) {
+            case NOT_READY:
+                mHrmState = HRMSensorState.READY;
+                break;
             case READY:
                 if (event.values[0] > 0) {
                     // The HRM sensor is ready to start processing
@@ -207,31 +257,29 @@ public class UserAuthActivity extends ActionBarActivity implements SensorEventLi
                     mUserAuthMsg.setText(R.string.user_auth_state_process_msg);
                     mHrmState = HRMSensorState.PROCESS;
                     this.timestamp = System.currentTimeMillis();
+                    delayedStopHRM();
                 }
                 break;
             case PROCESS:
-                long i = (System.currentTimeMillis() - this.timestamp) / 1000;
+                double i = (double)(System.currentTimeMillis() - this.timestamp) / 1000;
                 Log.i("HRM", "add point (" + i + ", " + event.values[0] + ")");
                 String val = Double.toString(event.values[0]);
                 mHRMProfile.add(val);
                 Log.i("HRM", "str = " + val);
                 mSeries.appendData(new DataPoint(i, event.values[0]), false, 500);
                 // Finish measurements in 10 seconds
-                delayedStopHRM();
                 break;
             case SUCCESS:
                 // Unregister HRM sensor listener
                 mSensorManager.unregisterListener(this);
                 mUserAuthState.setText(R.string.user_auth_state_success);
                 mUserAuthMsg.setText(R.string.user_auth_state_success_msg);
-                //saveHRMProfile();
-                //logHRMProfile();
                 break;
             case FAIL:
+                mSensorManager.unregisterListener(this);
                 mUserAuthState.setText(R.string.user_auth_state_fail);
                 mUserAuthMsg.setText(R.string.user_auth_state_fail_msg);
-                //saveHRMProfile();
-                //logHRMProfile();
+                delayedRestartHRM();
                 break;
         }
     }
@@ -252,10 +300,13 @@ public class UserAuthActivity extends ActionBarActivity implements SensorEventLi
         mHRMProfile.clear();
     }
 
-    private void saveHRMProfile() {
-        ArrayList<String> profile = getHRMSeries();
+    private void saveHRMProfile(boolean overwrite) {
+        // Check if HRMProfile already exists
+        if (!overwrite && getHRMProfile() != null) return;
+
         SharedPreferences sharedPref = PreferenceManager.getDefaultSharedPreferences(getApplicationContext());
         SharedPreferences.Editor editor = sharedPref.edit();
+        ArrayList<String> profile = getHRMSeries();
         editor.clear();
         editor.putString(getResources().getText(R.string.user_auth_bpm_profile).toString(),
                         serializeHRMProfile(profile));
@@ -281,7 +332,11 @@ public class UserAuthActivity extends ActionBarActivity implements SensorEventLi
         SharedPreferences sharedPref = PreferenceManager.getDefaultSharedPreferences(getApplicationContext());
         String hrmProfile = sharedPref.getString(getResources().getText(R.string.user_auth_bpm_profile).toString(), new String());
 
-        return deserializeHRMProfile(hrmProfile);
+        if (!hrmProfile.isEmpty()) {
+            return deserializeHRMProfile(hrmProfile);
+        }
+
+        return null;
     }
 
     private void logHRMProfile() {
@@ -292,4 +347,28 @@ public class UserAuthActivity extends ActionBarActivity implements SensorEventLi
         }
     }
 
+    private void displayConfidence(double confidence) {
+        TextView txtConf = (TextView) findViewById(R.id.user_auth_confidence);
+        txtConf.setText("Authentication Confidence Level: " + Math.round(confidence * 100) + "%");
+    }
+
+    private void clearAll() {
+
+        // Clear GraphView
+        mHrmState = HRMSensorState.NOT_READY;
+
+        // Clear HRM Profile
+        clearHRMProfile();
+        mSeries.resetData(new DataPoint[]{});
+        //mGraphView.removeAllSeries();
+
+        // Reset text views
+        mUserAuthState.setText(R.string.user_auth_state_ready);
+        mUserAuthMsg.setText(R.string.user_auth_ready_msg);
+        TextView txtConf = (TextView) findViewById(R.id.user_auth_confidence);
+        txtConf.setText("");
+
+        // Reset timestamp
+        timestamp = 0;
+    }
 }
